+++ src/brouwers/shop/payments/sisow/exceptions.py
@@ -1,0 +1,2 @@
+class InvalidIssuerURL(Exception):
+    pass
