+++ src/brouwers/shop/payments/tests/test_sisow.py
@@ -1,7 +1,8 @@
 from django.test import SimpleTestCase, TestCase
 
-from ...models import ShopConfiguration
-from ..sisow import CallbackForm, calculate_sha1
+from ...models import Payment, ShopConfiguration
+from ..sisow.api import calculate_ideal_sha1
+from ..sisow.forms import CallbackForm
 
 
 class Sha1Tests(SimpleTestCase):
@@ -12,7 +13,7 @@ class Sha1Tests(SimpleTestCase):
         purchaseid = "123"
         amount = "100"
 
-        sha1 = calculate_sha1(
+        sha1 = calculate_ideal_sha1(
             purchaseid=purchaseid,
             merchantid=merchantid,
             merchantkey=merchantkey,
@@ -33,13 +34,19 @@ class CallbackFormTests(TestCase):
         config.sisow_merchant_key = "28f31a03f4d272bb5d6dd6a345cce93b670e2f79"
         config.save()
 
+        cls.payment = Payment(data={
+            "sisow_transaction_request": {
+                "trxid": "TEST080536811624",
+            }
+        })
+
     def test_valid_callback_form(self):
         form = CallbackForm(data={
             "trxid": "TEST080536811624",
             "ec": "4cb92ef9-0bcc-4a",
             "status": "Success",
             "sha1": "54f2ae81eaa39fd8463f0f780a228fc3562faa43",
-        })
+        }, payment=self.payment)
 
         self.assertTrue(form.is_valid())
 
@@ -49,6 +56,6 @@ class CallbackFormTests(TestCase):
             "ec": "4cb92ef9-0bcc-4a",
             "status": "Success",
             "sha1": "ad534db9faba53ac064e2140a87429334e855d98",  # bad hash
-        })
+        }, payment=self.payment)
 
         self.assertFalse(form.is_valid())
