+++ src/brouwers/shop/urls.py
@@ -1,6 +1,7 @@
 from django.conf.urls import url
 
-from .debug_views import IdealPaymentView, PaymentCallbackView, PaymentView
+from .debug_views import IdealPaymentView, PaymentView
+from .payments.sisow.views import PaymentCallbackView
 from .views import (
     CartDetailView, CategoryDetailView, IndexView, ProductDetailView
 )
@@ -13,7 +14,10 @@ urlpatterns = [
     url(r'^products/(?P<slug>[-_\w]+)/', ProductDetailView.as_view(), name='product-detail'),
     url(r'^cart/(?P<pk>\d+)/', CartDetailView.as_view(), name='cart-detail'),
 
+    # payments
+    url(r'^payment/(?P<pk>[0-9]+)/callback/$', PaymentCallbackView.as_view(), name='sisow-payment-callback'),
+
+    # debug helpers
     url(r'^pay/$', PaymentView.as_view(), name='pay'),
     url(r'^pay/ideal/$', IdealPaymentView.as_view(), name='ideal-bank'),
-    url(r'^payment/callback/$', PaymentCallbackView.as_view(), name='payment-callback')
 ]
